Upgraded the bingo website
Added so shadow balls appear when you enter the website so you don't have weird gabs, made the button bigger for easy pressing and the text of it. Made it so the ball that gets picked gets put in the right index. And I added a logo to the website.